finish deletal of aglib.heap.heap , get rid of aglib.heap.__init__.py as well
